klikinsäästäjä/pyproject.toml
[tool.poetry]
name = "klikinsaastaja-ng"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "AGPL"
packages = [
    { include = "klikinsaastaja_ng", from = "src/" },
]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.poetry.dependencies]
python = "^3.10"
mwparserfromhell = "^0.6.6"
wikipedia = "^1.4.0"
newspaper4k = "^0.9.2"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
EdgeGPT-fork = "^0.14.1"

requests-cache = {version = "^1.1.1", optional = true}
openai = {version = "^1.11.0", optional = true}
sentence-transformers = {version = "^2.2.2", optional = true}
Jinja2 = "^3.1.3"

[tool.poetry.dev-dependencies]
black = "^24.1.1"
ruff = "^0.2.0"
gradio = "^4.16.0"
poetry = "^1.7.1"
isort = "^5.13.2"
ipykernel = "^6.29.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

